Разделение SpendExtension.class
