+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/annotation/Spend.java
@@ -1,5 +1,7 @@
-package guru.qa.niffler.jupiter;
+package guru.qa.niffler.jupiter.annotation;
 
+import guru.qa.niffler.jupiter.extension.SpendExtension;
+import guru.qa.niffler.jupiter.extension.SpendExtensionParameterResolver;
 import guru.qa.niffler.model.CurrencyValues;
 import org.junit.jupiter.api.extension.ExtendWith;
 
@@ -10,7 +12,7 @@ import java.lang.annotation.Target;
 
 @Target(ElementType.METHOD)
 @Retention(RetentionPolicy.RUNTIME)
-@ExtendWith(SpendExtension.class)
+@ExtendWith({SpendExtension.class, SpendExtensionParameterResolver.class})
 public @interface Spend {
 
     String username();
