+++ niffler-e-2-e-tests/src/test/java/guru/qa/niffler/jupiter/extension/SpendExtensionParameterResolver.java
@@ -1,0 +1,24 @@
+package guru.qa.niffler.jupiter.extension;
+
+import guru.qa.niffler.model.SpendJson;
+import org.junit.jupiter.api.extension.ExtensionContext;
+import org.junit.jupiter.api.extension.ParameterContext;
+import org.junit.jupiter.api.extension.ParameterResolutionException;
+import org.junit.jupiter.api.extension.ParameterResolver;
+
+import static guru.qa.niffler.jupiter.extension.SpendExtension.NAMESPACE;
+
+public class SpendExtensionParameterResolver implements ParameterResolver {
+
+    @Override
+    public boolean supportsParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
+        return parameterContext.getParameter()
+                .getType()
+                .isAssignableFrom(SpendJson.class);
+    }
+
+    @Override
+    public SpendJson resolveParameter(ParameterContext parameterContext, ExtensionContext extensionContext) throws ParameterResolutionException {
+        return extensionContext.getStore(NAMESPACE).get("spend", SpendJson.class);
+    }
+}
